Fix bug by passing a function to setChatLog
Previously, the value of chatLog to which we appended incoming
messages was bound when useEffect was first called. This meant
that all messages were appended to the initial (empty) chat log,
so they'd appear to replace whatever messages were already there.

Now we get the most recent chat log via the callback passed to
setChatLog.